fix: safe HostAggregate lifespan/downtime and uint64 max boot (ask 34)
Guard Lifespan and Downtime against uint64 underflow when LastSeen < FirstBoot
or Uptime exceeds lifespan. Track per-host max BootTime with uint64 in
LoadAggregates to match file aggregation and avoid int64 conversion bugs.

Tests: lifespan underflow, downtime edge cases, LastKernel near MaxInt64.
Made-with: Cursor

+++ internal/goprecords/db.go
@@ -42,12 +42,12 @@ func LoadAggregates(ctx context.Context, db *sql.DB) (*Aggregates, error) {
 		KernelMajor: make(map[string]*Aggregate),
 		KernelName:  make(map[string]*Aggregate),
 	}
-	hostMaxBoot := make(map[string]int64)
+	hostMaxBoot := make(map[string]uint64)
 	hostLastKernel := make(map[string]string)
 
 	for _, rec := range records {
-		if rec.BootTime >= uint64(hostMaxBoot[rec.Host]) {
-			hostMaxBoot[rec.Host] = int64(rec.BootTime)
+		if rec.BootTime >= hostMaxBoot[rec.Host] {
+			hostMaxBoot[rec.Host] = rec.BootTime
 			hostLastKernel[rec.Host] = rec.OS
 		}
 		if _, ok := out.Host[rec.Host]; !ok {
